refactor: move renderTree to its own file
